+++ Text/Megaparsec.hs
@@ -101,6 +101,14 @@ module Text.Megaparsec
   , region
   , takeRest
   , atEnd
+  , char
+  , anyChar
+  , notChar
+  , oneOf
+  , noneOf
+  , satisfy
+  , string
+  , string'
     -- * Parser state combinators
   , getInput
   , setInput
@@ -129,6 +137,7 @@ import Control.Monad.State.Class hiding (state)
 import Control.Monad.Trans
 import Control.Monad.Trans.Identity
 import Data.Data (Data)
+import Data.Function (on)
 import Data.List.NonEmpty (NonEmpty (..))
 import Data.Maybe (fromJust)
 import Data.Proxy
@@ -150,11 +159,14 @@ import qualified Control.Monad.Trans.State.Lazy    as L
 import qualified Control.Monad.Trans.State.Strict  as S
 import qualified Control.Monad.Trans.Writer.Lazy   as L
 import qualified Control.Monad.Trans.Writer.Strict as S
+import qualified Data.CaseInsensitive as CI
 import qualified Data.List.NonEmpty                as NE
 import qualified Data.Set                          as E
 
 #if !MIN_VERSION_base(4,8,0)
 import Control.Applicative
+import Data.Foldable (Foldable (), elem, notElem)
+import Prelude hiding (elem, notElem)
 #endif
 
 -- $reexports
@@ -1408,6 +1420,118 @@ takeRest = takeWhileP Nothing (const True)
 atEnd :: MonadParsec e s m => m Bool
 atEnd = option False (True <$ eof)
 {-# INLINE atEnd #-}
+
+-- | @'char' c@ parses a single character @c@.
+--
+-- > semicolon = char ';'
+
+char :: MonadParsec e s m => Token s -> m (Token s)
+char c = token testChar (Just c)
+  where
+    f x = Tokens (x:|[])
+    testChar x =
+      if x == c
+        then Right x
+        else Left (pure (f x), E.singleton (f c))
+{-# INLINE char #-}
+
+-- | This parser succeeds for any character. Returns the parsed character.
+
+anyChar :: MonadParsec e s m => m (Token s)
+anyChar = satisfy (const True) <?> "character"
+{-# INLINE anyChar #-}
+
+-- | Match any character but the given one. It's a good idea to attach a
+-- 'label' to this parser manually.
+--
+-- @since 6.0.0
+
+notChar :: MonadParsec e s m => Token s -> m (Token s)
+notChar c = satisfy (/= c)
+{-# INLINE notChar #-}
+
+-- | @'oneOf' cs@ succeeds if the current character is in the supplied
+-- collection of characters @cs@. Returns the parsed character. Note that
+-- this parser cannot automatically generate the “expected” component of
+-- error message, so usually you should label it manually with 'label' or
+-- ('<?>').
+--
+-- See also: 'satisfy'.
+--
+-- > digit = oneOf ['0'..'9'] <?> "digit"
+--
+-- __Performance note__: prefer 'satisfy' when you can because it's faster
+-- when you have only a couple of tokens to compare to:
+--
+-- > quoteFast = satisfy (\x -> x == '\'' || x == '\"')
+-- > quoteSlow = oneOf "'\""
+
+oneOf :: (Foldable f, MonadParsec e s m)
+  => f (Token s)
+  -> m (Token s)
+oneOf cs = satisfy (`elem` cs)
+{-# INLINE oneOf #-}
+
+-- | As the dual of 'oneOf', @'noneOf' cs@ succeeds if the current character
+-- /not/ in the supplied list of characters @cs@. Returns the parsed
+-- character. Note that this parser cannot automatically generate the
+-- “expected” component of error message, so usually you should label it
+-- manually with 'label' or ('<?>').
+--
+-- See also: 'satisfy'.
+--
+-- __Performance note__: prefer 'satisfy' and 'notChar' when you can because
+-- it's faster.
+
+noneOf :: (Foldable f, MonadParsec e s m)
+  => f (Token s)
+  -> m (Token s)
+noneOf cs = satisfy (`notElem` cs)
+{-# INLINE noneOf #-}
+
+-- | The parser @'satisfy' f@ succeeds for any character for which the
+-- supplied function @f@ returns 'True'. Returns the character that is
+-- actually parsed.
+--
+-- > digitChar = satisfy isDigit <?> "digit"
+-- > oneOf cs  = satisfy (`elem` cs)
+
+satisfy :: MonadParsec e s m
+  => (Token s -> Bool) -- ^ Predicate to apply
+  -> m (Token s)
+satisfy f = token testChar Nothing
+  where
+    testChar x =
+      if f x
+        then Right x
+        else Left (pure (Tokens (x:|[])), E.empty)
+{-# INLINE satisfy #-}
+
+----------------------------------------------------------------------------
+-- Sequence of characters
+
+-- | @'string' s@ parses a sequence of characters given by @s@. Returns the
+-- parsed string (i.e. @s@).
+--
+-- > divOrMod = string "div" <|> string "mod"
+
+string :: MonadParsec e s m
+  => Tokens s
+  -> m (Tokens s)
+string = tokens (==)
+{-# INLINE string #-}
+
+-- | The same as 'string', but case-insensitive. On success returns string
+-- cased as actually parsed input.
+--
+-- >>> parseTest (string' "foobar") "foObAr"
+-- "foObAr"
+
+string' :: (MonadParsec e s m, CI.FoldCase (Tokens s))
+  => Tokens s
+  -> m (Tokens s)
+string' = tokens ((==) `on` CI.mk)
+{-# INLINE string' #-}
 
 ----------------------------------------------------------------------------
 -- Parser state combinators
